% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airport}
\alias{airport}
\title{US Air Traffic Network}
\format{
igraph object
}
\usage{
airport
}
\description{
A weighted and directed network of passenger air traffic in the United States in 2019.
Each edge represents a single takeoff and landing, and therefore does not consider possible layovers,
connecting flights, round trips, etc. This is the directed version of the undirected air traffic network
used by Neal (2022) to illustrate \code{backbone::disparity()}. GRAND attributes have already been added
using \code{\link[=grand]{grand()}}.
}
\references{
Neal, Z. P. (2022). backbone: An R Package to Extract Network Backbones. \emph{PLOS ONE, 17}, e0269137. \doi{10.1371/journal.pone.0269137}
}
\keyword{datasets}
