% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosponsor}
\alias{cosponsor}
\title{US Senate Co-Sponsorship Network}
\format{
igraph object
}
\usage{
cosponsor
}
\description{
A bipartite network representing US Senators' (co-)sponsorship of Senate Bills during the 116th
session (2019-2020). It was obtained using \code{incidentally::incidence.from.congress()} following the
procedure described by Neal (2022). GRAND attributes have already been added using \code{\link[=grand]{grand()}}.
}
\references{
Neal, Z. P. (2022). Constructing legislative networks in R using incidentally and backbone. \emph{Connections, 42}, 1-9. \doi{10.2478/connections-2019.026}
}
\keyword{datasets}
