% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grand.text.R
\name{grand.text}
\alias{grand.text}
\title{Generate a Guidelines for Reporting About Network Data (GRAND) narrative summary}
\usage{
grand.text(G, digits = 3)
}
\arguments{
\item{G}{An \code{\link{igraph}} object with GRAND attributed}

\item{digits}{numeric: number of decimal places to report}
}
\value{
string: Narrative summary of G
}
\description{
Generate a Guidelines for Reporting About Network Data (GRAND) narrative summary
}
\examples{
data(airport)  #Load example data
grand.text(airport)  #Generate narrative
}
