% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scan2}
\alias{scan2}
\title{Restricts \code{scan()} input to a specified format}
\usage{
scan2(prompt, type)
}
\arguments{
\item{prompt}{string: prompt for user input}

\item{type}{string: required format for input}
}
\value{
user input in specified format
}
\description{
Restricts \code{scan()} input to a specified format
}
\examples{
character <- scan2(prompt = "Type any character", type = "character")
numeric <- scan2(prompt = "Type any number", type = "numeric")
integer <- scan2(prompt = "Type any number", type = "integer")
custom <- scan2(prompt = "Yes or No?", type = c("Y","N"))
}
