% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senate}
\alias{senate}
\title{US Senate Network}
\format{
igraph object
}
\usage{
senate
}
\description{
A signed network representing US Senators' alliances and antagonisms, inferred from
\code{\link[=cosponsor]{cosponsor()}} using \code{backbone::sdsm()} following the procedure described by Neal (2022).
GRAND attributes have already been added using \code{\link[=grand]{grand()}}.
}
\references{
Neal, Z. P. (2022). Constructing legislative networks in R using incidentally and backbone. \emph{Connections, 42}, 1-9. \doi{10.2478/connections-2019.026}
}
\keyword{datasets}
