% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{menu2}
\alias{menu2}
\title{Returns \code{menu()} response as choice text}
\usage{
menu2(choices, title, loop = FALSE)
}
\arguments{
\item{choices}{a character vector of choices}

\item{title}{a character string to be used as the title of the menu. \code{NULL} is also accepted.}

\item{loop}{boolean: should the menu loop to allow multiple choices?}
}
\value{
string: the chosen option
}
\description{
Returns \code{menu()} response as choice text
}
\examples{
choice <- menu2(choices = c("A", "B", "C"), title = "Choose an option", loop = TRUE)
}
